Create shared create-team command
